"6:20pm.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 332 and 668 out of 1000.
The time was 00:00:04.7681761 for: iteration 0
Starting timing for: iteration 1
Winrate is 469 and 531 out of 1000.
The time was 00:00:05.5162523 for: iteration 1
Starting timing for: iteration 2
Winrate is 589 and 411 out of 1000.
The time was 00:00:06.0338665 for: iteration 2
Starting timing for: iteration 3
Winrate is 622 and 378 out of 1000.
The time was 00:00:06.2311515 for: iteration 3
Starting timing for: iteration 4
Winrate is 664 and 336 out of 1000.
The time was 00:00:06.6690796 for: iteration 4
Starting timing for: iteration 5
Winrate is 673 and 327 out of 1000.
The time was 00:00:06.5804610 for: iteration 5
Starting timing for: iteration 6
Winrate is 685 and 315 out of 1000.
The time was 00:00:06.6751601 for: iteration 6
Starting timing for: iteration 7
Winrate is 733 and 267 out of 1000.
The time was 00:00:07.0719601 for: iteration 7
Starting timing for: iteration 8
Winrate is 717 and 283 out of 1000.
The time was 00:00:06.8585667 for: iteration 8
Starting timing for: iteration 9
Winrate is 712 and 288 out of 1000.
The time was 00:00:07.2856222 for: iteration 9
Starting timing for: iteration 10
Winrate is 741 and 259 out of 1000.
The time was 00:00:06.8837203 for: iteration 10
Starting timing for: iteration 11
Winrate is 716 and 284 out of 1000.
The time was 00:00:06.5479149 for: iteration 11
Starting timing for: iteration 12
Winrate is 725 and 275 out of 1000.
The time was 00:00:06.4836237 for: iteration 12
Starting timing for: iteration 13
Winrate is 702 and 298 out of 1000.
The time was 00:00:06.4902534 for: iteration 13
Starting timing for: iteration 14
Winrate is 728 and 272 out of 1000.
The time was 00:00:06.5035889 for: iteration 14
```

It seems KL can do it, but the lesson from this is that poorly tuned initial range can really hurt it.

6:40pm.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 328 and 672 out of 1000.
The time was 00:00:04.7991386 for: iteration 0
Starting timing for: iteration 1
Winrate is 594 and 406 out of 1000.
The time was 00:00:06.1572238 for: iteration 1
Starting timing for: iteration 2
Winrate is 683 and 317 out of 1000.
The time was 00:00:06.1791723 for: iteration 2
Starting timing for: iteration 3
Winrate is 694 and 306 out of 1000.
The time was 00:00:06.2670857 for: iteration 3
Starting timing for: iteration 4
Winrate is 709 and 291 out of 1000.
The time was 00:00:06.2621960 for: iteration 4
Starting timing for: iteration 5
Winrate is 714 and 286 out of 1000.
The time was 00:00:06.1706688 for: iteration 5
Starting timing for: iteration 6
Winrate is 710 and 290 out of 1000.
The time was 00:00:06.6819573 for: iteration 6
Starting timing for: iteration 7
Winrate is 744 and 256 out of 1000.
The time was 00:00:06.2841767 for: iteration 7
Starting timing for: iteration 8
Winrate is 714 and 286 out of 1000.
The time was 00:00:06.6550492 for: iteration 8
Starting timing for: iteration 9
Winrate is 736 and 264 out of 1000.
The time was 00:00:07.2354732 for: iteration 9
Starting timing for: iteration 10
Winrate is 730 and 270 out of 1000.
The time was 00:00:06.7043874 for: iteration 10
Starting timing for: iteration 11
CuSolve function failed!
The leading minor of order 1 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:720: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

7:10pm. Done with lunch. I am not going to be doing any more programming as it is damn late. Note that up to the point where it blows up, it seems that KL divergence is working quite well.

```
inm {scale scale_r scale'} = module_map (inl _ {eta upper mid} -> eta + one) {scale scale_r scale'}
```

It seems that plopping the scale directly does wonders. I think this performance is on par with squared error now, but I need to figure out what is wrong with the thing first.

Let me turn off KFAC.

Ok...I will be doing more programming."