"7:20pm.
```
Starting timing for: iteration 134
Winrate is 78 and 22 out of 100.
The time was 00:00:00.6439018 for: iteration 134
Starting timing for: iteration 135
Winrate is 70 and 30 out of 100.
The time was 00:00:00.6338366 for: iteration 135
Starting timing for: iteration 136
Winrate is 65 and 35 out of 100.
The time was 00:00:00.5563605 for: iteration 136
Starting timing for: iteration 137
Winrate is 71 and 29 out of 100.
The time was 00:00:00.5832485 for: iteration 137
Starting timing for: iteration 138
CuSolve function failed!
The leading minor of order 7 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:720: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

Just why is this happening?

7:25pm. I have no idea. I am not going to be answering this today anyway.

At this point, it is time to rest. At the very least, it does seem like learnable eligibility traces are a win. If I could deal with the KL cost, that would be great. To be honest, I expected it would work hugely better than squared error and here it is giving me trouble. And reward rescaling seems to be a complete bust."