"8pm.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 288 and 712 out of 1000.
The time was 00:00:03.3432023 for: iteration 0
Starting timing for: iteration 1
Winrate is 347 and 653 out of 1000.
The time was 00:00:03.6313491 for: iteration 1
Starting timing for: iteration 2
Winrate is 436 and 564 out of 1000.
The time was 00:00:04.5647886 for: iteration 2
Starting timing for: iteration 3
Winrate is 535 and 465 out of 1000.
The time was 00:00:05.2818038 for: iteration 3
Starting timing for: iteration 4
Winrate is 556 and 444 out of 1000.
The time was 00:00:04.9300997 for: iteration 4
Starting timing for: iteration 5
Winrate is 573 and 427 out of 1000.
The time was 00:00:05.0161797 for: iteration 5
Starting timing for: iteration 6
Winrate is 611 and 389 out of 1000.
The time was 00:00:04.9099209 for: iteration 6
Starting timing for: iteration 7
Winrate is 618 and 382 out of 1000.
The time was 00:00:05.1200535 for: iteration 7
Starting timing for: iteration 8
Winrate is 571 and 429 out of 1000.
The time was 00:00:05.4800081 for: iteration 8
Starting timing for: iteration 9
Winrate is 604 and 396 out of 1000.
The time was 00:00:05.5764720 for: iteration 9
Starting timing for: iteration 10
Winrate is 622 and 378 out of 1000.
The time was 00:00:05.2536837 for: iteration 10
Starting timing for: iteration 11
Winrate is 630 and 370 out of 1000.
The time was 00:00:05.7187975 for: iteration 11
Starting timing for: iteration 12
Winrate is 598 and 402 out of 1000.
The time was 00:00:05.7951640 for: iteration 12
Starting timing for: iteration 13
Winrate is 643 and 357 out of 1000.
The time was 00:00:05.6773009 for: iteration 13
Starting timing for: iteration 14
Winrate is 650 and 350 out of 1000.
The time was 00:00:05.4105547 for: iteration 14
```

Oh, rescaling the reward does make it run in fact.

8:05pm.

```
The learning_rate is 2 ** -12
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 301 and 699 out of 1000.
The time was 00:00:02.9189292 for: iteration 0
Starting timing for: iteration 1
Winrate is 269 and 731 out of 1000.
The time was 00:00:02.7739973 for: iteration 1
Starting timing for: iteration 2
Winrate is 285 and 715 out of 1000.
The time was 00:00:02.7280585 for: iteration 2
Starting timing for: iteration 3
Winrate is 316 and 684 out of 1000.
The time was 00:00:02.6423525 for: iteration 3
Starting timing for: iteration 4
Winrate is 270 and 730 out of 1000.
The time was 00:00:02.7170872 for: iteration 4
Starting timing for: iteration 5
Winrate is 274 and 726 out of 1000.
The time was 00:00:02.6987454 for: iteration 5
Starting timing for: iteration 6
```

The above is squared error.

Pretty much everything I try is unfailingly bad with the * 10 reward. But / 10 seems to help quite a bit. Let me try / 100."