"8:20pm.
```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 468 and 532 out of 1000.
The time was 00:00:05.9220331 for: iteration 0
Starting timing for: iteration 1
Winrate is 691 and 309 out of 1000.
The time was 00:00:06.5743564 for: iteration 1
Starting timing for: iteration 2
Winrate is 722 and 278 out of 1000.
The time was 00:00:06.6717263 for: iteration 2
Starting timing for: iteration 3
Winrate is 717 and 283 out of 1000.
The time was 00:00:06.4375027 for: iteration 3
Starting timing for: iteration 4
Winrate is 720 and 280 out of 1000.
The time was 00:00:06.5385565 for: iteration 4
Starting timing for: iteration 5
Winrate is 719 and 281 out of 1000.
The time was 00:00:06.6918237 for: iteration 5
Starting timing for: iteration 6
Winrate is 717 and 283 out of 1000.
The time was 00:00:07.0471710 for: iteration 6
Starting timing for: iteration 7
Winrate is 756 and 244 out of 1000.
The time was 00:00:06.7283536 for: iteration 7
Starting timing for: iteration 8
Winrate is 711 and 289 out of 1000.
The time was 00:00:06.6707580 for: iteration 8
Starting timing for: iteration 9
Winrate is 722 and 278 out of 1000.
The time was 00:00:06.9608058 for: iteration 9
Starting timing for: iteration 10
Winrate is 741 and 259 out of 1000.
The time was 00:00:06.7199838 for: iteration 10
Starting timing for: iteration 11
Winrate is 746 and 254 out of 1000.
The time was 00:00:06.5899890 for: iteration 11
Starting timing for: iteration 12
Winrate is 718 and 282 out of 1000.
The time was 00:00:06.6788855 for: iteration 12
Starting timing for: iteration 13
Winrate is 700 and 300 out of 1000.
The time was 00:00:06.4302047 for: iteration 13
Starting timing for: iteration 14
Winrate is 736 and 264 out of 1000.
The time was 00:00:06.5813008 for: iteration 14
```

Dividing the reward by 10 works wonders. This is with squared error.

Seriously, fuck deep learning.

8:25pm.

```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 464 and 536 out of 1000.
The time was 00:00:05.0844328 for: iteration 0
Starting timing for: iteration 1
Winrate is 661 and 339 out of 1000.
The time was 00:00:06.1547819 for: iteration 1
Starting timing for: iteration 2
Winrate is 695 and 305 out of 1000.
The time was 00:00:06.4970377 for: iteration 2
Starting timing for: iteration 3
Winrate is 692 and 308 out of 1000.
The time was 00:00:06.4020696 for: iteration 3
Starting timing for: iteration 4
Winrate is 699 and 301 out of 1000.
The time was 00:00:06.4838336 for: iteration 4
Starting timing for: iteration 5
Winrate is 735 and 265 out of 1000.
The time was 00:00:06.2132330 for: iteration 5
Starting timing for: iteration 6
Winrate is 679 and 321 out of 1000.
The time was 00:00:06.7318722 for: iteration 6
Starting timing for: iteration 7
Winrate is 735 and 265 out of 1000.
The time was 00:00:06.1819911 for: iteration 7
Starting timing for: iteration 8
Winrate is 725 and 275 out of 1000.
The time was 00:00:06.5259088 for: iteration 8
Starting timing for: iteration 9
Winrate is 688 and 312 out of 1000.
The time was 00:00:06.5105071 for: iteration 9
Starting timing for: iteration 10
Winrate is 733 and 267 out of 1000.
The time was 00:00:06.6026312 for: iteration 10
Starting timing for: iteration 11
Winrate is 723 and 277 out of 1000.
The time was 00:00:06.1561540 for: iteration 11
Starting timing for: iteration 12
Winrate is 718 and 282 out of 1000.
The time was 00:00:06.4343624 for: iteration 12
Starting timing for: iteration 13
CuSolve function failed!
The leading minor of order 11 is not positive definite.
C:/Users/Marko/Source/Repos/The Spiral Language/Temporary/bin/Release/cuda_kernels.cu:720: block: [0,0,0], thread: [0,0,0] Assertion `is_info_zero` failed.
```

I give up. There is no reason at all to use the Gaussian cost.

At least eligibility traces seem like a win.

Let me test TD 0 vs MC and then I will really stop for the day."