"9:50am.
```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 433 and 567 out of 1000.
The time was 00:00:05.9544956 for: iteration 0
Starting timing for: iteration 1
Winrate is 482 and 518 out of 1000.
The time was 00:00:06.1141938 for: iteration 1
Starting timing for: iteration 2
Winrate is 564 and 436 out of 1000.
The time was 00:00:07.0105648 for: iteration 2
Starting timing for: iteration 3
Winrate is 574 and 426 out of 1000.
The time was 00:00:06.8765571 for: iteration 3
Starting timing for: iteration 4
Winrate is 644 and 356 out of 1000.
The time was 00:00:07.1468540 for: iteration 4
Starting timing for: iteration 5
Winrate is 676 and 324 out of 1000.
The time was 00:00:07.3748297 for: iteration 5
Starting timing for: iteration 6
Winrate is 730 and 270 out of 1000.
The time was 00:00:07.6184074 for: iteration 6
Starting timing for: iteration 7
Winrate is 727 and 273 out of 1000.
The time was 00:00:07.2009327 for: iteration 7
Starting timing for: iteration 8
Winrate is 722 and 278 out of 1000.
The time was 00:00:07.1296153 for: iteration 8
Starting timing for: iteration 9
Winrate is 718 and 282 out of 1000.
The time was 00:00:07.0878195 for: iteration 9
Starting timing for: iteration 10
Winrate is 710 and 290 out of 1000.
The time was 00:00:06.7575055 for: iteration 10
Starting timing for: iteration 11
Winrate is 711 and 289 out of 1000.
The time was 00:00:06.4852689 for: iteration 11
Starting timing for: iteration 12
Winrate is 708 and 292 out of 1000.
The time was 00:00:06.5087538 for: iteration 12
Starting timing for: iteration 13
Winrate is 703 and 297 out of 1000.
The time was 00:00:06.4854094 for: iteration 13
Starting timing for: iteration 14
Winrate is 689 and 311 out of 1000.
The time was 00:00:07.0217333 for: iteration 14
```

In truth, I would be perfectly fine with optimizing the mean directly if it would mean not passing the oversized gradients to the bottom layers.

9:55am.

```
inm _ = sqr error * primal scale |> as_cost
```

Let me try it like this.

```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 323 and 677 out of 1000.
The time was 00:00:04.7490433 for: iteration 0
Starting timing for: iteration 1
Winrate is 542 and 458 out of 1000.
The time was 00:00:06.6201825 for: iteration 1
Starting timing for: iteration 2
Winrate is 720 and 280 out of 1000.
The time was 00:00:07.2562028 for: iteration 2
Starting timing for: iteration 3
Winrate is 506 and 494 out of 1000.
The time was 00:00:04.9611067 for: iteration 3
Starting timing for: iteration 4
Winrate is 540 and 460 out of 1000.
The time was 00:00:05.4758298 for: iteration 4
Starting timing for: iteration 5
Winrate is 584 and 416 out of 1000.
The time was 00:00:07.4925095 for: iteration 5
Starting timing for: iteration 6
Winrate is 648 and 352 out of 1000.
The time was 00:00:08.0491445 for: iteration 6
Starting timing for: iteration 7
```

It is unstable.

Let me raise the LR once more and then I will remove the scale from the mean cost.

Just forget it. It never made sense in the first place. If I want to modify the gradients specifically then, I am going to need a different mechanism.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 366 and 634 out of 1000.
The time was 00:00:04.9817330 for: iteration 0
Starting timing for: iteration 1
Winrate is 577 and 423 out of 1000.
The time was 00:00:06.3008773 for: iteration 1
Starting timing for: iteration 2
Winrate is 606 and 394 out of 1000.
The time was 00:00:06.6762812 for: iteration 2
Starting timing for: iteration 3
Winrate is 667 and 333 out of 1000.
The time was 00:00:07.0241784 for: iteration 3
Starting timing for: iteration 4
Winrate is 702 and 298 out of 1000.
The time was 00:00:07.3499967 for: iteration 4
Starting timing for: iteration 5
Winrate is 729 and 271 out of 1000.
The time was 00:00:07.3872548 for: iteration 5
Starting timing for: iteration 6
Winrate is 732 and 268 out of 1000.
The time was 00:00:07.4367997 for: iteration 6
Starting timing for: iteration 7
Winrate is 750 and 250 out of 1000.
The time was 00:00:07.9664736 for: iteration 7
Starting timing for: iteration 8
Winrate is 717 and 283 out of 1000.
The time was 00:00:07.4966256 for: iteration 8
Starting timing for: iteration 9
Winrate is 715 and 285 out of 1000.
The time was 00:00:07.4853583 for: iteration 9
Starting timing for: iteration 10
Winrate is 725 and 275 out of 1000.
The time was 00:00:07.0738918 for: iteration 10
Starting timing for: iteration 11
Winrate is 748 and 252 out of 1000.
The time was 00:00:07.4424847 for: iteration 11
Starting timing for: iteration 12
Winrate is 722 and 278 out of 1000.
The time was 00:00:07.3889559 for: iteration 12
Starting timing for: iteration 13
Winrate is 723 and 277 out of 1000.
The time was 00:00:07.1323859 for: iteration 13
Starting timing for: iteration 14
Winrate is 737 and 263 out of 1000.
The time was 00:00:07.0693914 for: iteration 14
```

The performance of this is not bad at all.

Let me try multiplying the rewards by 100 and eta by -100.

```
inm _ = sqr error * primal scale |> as_cost
```

Before that, I really do think that this multiplication should be done inside after all."