"10:35am.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 360 and 640 out of 1000.
The time was 00:00:06.2022695 for: iteration 0
Starting timing for: iteration 1
Winrate is 596 and 404 out of 1000.
The time was 00:00:06.7538973 for: iteration 1
Starting timing for: iteration 2
Winrate is 669 and 331 out of 1000.
The time was 00:00:07.3910594 for: iteration 2
Starting timing for: iteration 3
Winrate is 679 and 321 out of 1000.
The time was 00:00:07.4108909 for: iteration 3
Starting timing for: iteration 4
Winrate is 738 and 262 out of 1000.
The time was 00:00:06.8512736 for: iteration 4
Starting timing for: iteration 5
Winrate is 721 and 279 out of 1000.
The time was 00:00:07.2156228 for: iteration 5
Starting timing for: iteration 6
Winrate is 765 and 235 out of 1000.
The time was 00:00:06.7856951 for: iteration 6
Starting timing for: iteration 7
Winrate is 752 and 248 out of 1000.
The time was 00:00:06.7366652 for: iteration 7
Starting timing for: iteration 8
Winrate is 760 and 240 out of 1000.
The time was 00:00:07.2667920 for: iteration 8
Starting timing for: iteration 9
Winrate is 757 and 243 out of 1000.
The time was 00:00:06.9950312 for: iteration 9
Starting timing for: iteration 10
Winrate is 743 and 257 out of 1000.
The time was 00:00:06.7722159 for: iteration 10
Starting timing for: iteration 11
Winrate is 733 and 267 out of 1000.
The time was 00:00:06.6448016 for: iteration 11
Starting timing for: iteration 12
Winrate is 757 and 243 out of 1000.
The time was 00:00:06.8066010 for: iteration 12
Starting timing for: iteration 13
Winrate is 748 and 252 out of 1000.
The time was 00:00:06.7136538 for: iteration 13
Starting timing for: iteration 14
Winrate is 751 and 249 out of 1000.
The time was 00:00:06.4709797 for: iteration 14
```

Holy shit. This is probably the best run I've ever had. Reward rescaling works!

```
inm _ = sqr (error) |> as_cost
```

Ah, but I did not put this rescaling of error. So it only works up to the point of passing rescaled rewards to the actor.

10:45am. Well, there is a way I could normalize the two parts but it is a bit hacky. Let me just bit the bullet and do it."