"10:55am. I think before long, I'll forget what that run was. Well, I'll dig it out somehow.
```
        inm _ = abs (abs (primal error) - scale) / primal scale |> as_cost
        inm scale = max epsilon scale

        inm _ = sqr (error / primal scale) |> as_cost
```

Let ne try this again.

11am.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 288 and 712 out of 1000.
The time was 00:00:04.4464482 for: iteration 0
Starting timing for: iteration 1
Winrate is 469 and 531 out of 1000.
The time was 00:00:06.7235133 for: iteration 1
Starting timing for: iteration 2
Winrate is 544 and 456 out of 1000.
The time was 00:00:06.7413478 for: iteration 2
Starting timing for: iteration 3
Winrate is 611 and 389 out of 1000.
The time was 00:00:06.7410380 for: iteration 3
Starting timing for: iteration 4
Winrate is 656 and 344 out of 1000.
The time was 00:00:07.8958773 for: iteration 4
Starting timing for: iteration 5
Winrate is 668 and 332 out of 1000.
The time was 00:00:08.1074178 for: iteration 5
Starting timing for: iteration 6
Winrate is 701 and 299 out of 1000.
The time was 00:00:08.3964604 for: iteration 6
Starting timing for: iteration 7
Winrate is 706 and 294 out of 1000.
The time was 00:00:07.9226798 for: iteration 7
Starting timing for: iteration 8
Winrate is 705 and 295 out of 1000.
The time was 00:00:08.0298582 for: iteration 8
Starting timing for: iteration 9
Winrate is 717 and 283 out of 1000.
The time was 00:00:08.2254192 for: iteration 9
Starting timing for: iteration 10
Winrate is 711 and 289 out of 1000.
The time was 00:00:07.9210111 for: iteration 10
Starting timing for: iteration 11
Winrate is 723 and 277 out of 1000.
The time was 00:00:08.3422701 for: iteration 11
Starting timing for: iteration 12
Winrate is 729 and 271 out of 1000.
The time was 00:00:08.3175086 for: iteration 12
Starting timing for: iteration 13
Winrate is 723 and 277 out of 1000.
The time was 00:00:08.5014776 for: iteration 13
Starting timing for: iteration 14
Winrate is 746 and 254 out of 1000.
The time was 00:00:08.3204708 for: iteration 14
```

This might be a good idea. Let me push up the learning rates."