"11am.
```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 310 and 690 out of 1000.
The time was 00:00:03.5370771 for: iteration 0
Starting timing for: iteration 1
Winrate is 345 and 655 out of 1000.
The time was 00:00:04.8441757 for: iteration 1
Starting timing for: iteration 2
Winrate is 472 and 528 out of 1000.
The time was 00:00:06.0782651 for: iteration 2
Starting timing for: iteration 3
Winrate is 433 and 567 out of 1000.
The time was 00:00:04.6432235 for: iteration 3
Starting timing for: iteration 4
Winrate is 406 and 594 out of 1000.
The time was 00:00:03.1854369 for: iteration 4
Starting timing for: iteration 5
Winrate is 393 and 607 out of 1000.
The time was 00:00:03.1449522 for: iteration 5
Starting timing for: iteration 6
Winrate is 423 and 577 out of 1000.
The time was 00:00:03.3153669 for: iteration 6
Starting timing for: iteration 7
Winrate is 376 and 624 out of 1000.
The time was 00:00:03.3593626 for: iteration 7
Starting timing for: iteration 8
Winrate is 410 and 590 out of 1000.
The time was 00:00:03.2456302 for: iteration 8
Starting timing for: iteration 9
Winrate is 382 and 618 out of 1000.
The time was 00:00:03.1863742 for: iteration 9
Starting timing for: iteration 10
Winrate is 351 and 649 out of 1000.
The time was 00:00:03.1939470 for: iteration 10
Starting timing for: iteration 11
Winrate is 413 and 587 out of 1000.
The time was 00:00:03.1180248 for: iteration 11
Starting timing for: iteration 12
Winrate is 373 and 627 out of 1000.
The time was 00:00:03.0433093 for: iteration 12
Starting timing for: iteration 13
Winrate is 415 and 585 out of 1000.
The time was 00:00:03.3150540 for: iteration 13
Starting timing for: iteration 14
Winrate is 397 and 603 out of 1000.
The time was 00:00:03.2286886 for: iteration 14
```

Not what I thought it would be.

Maybe I should lower the LRs?

11:05am.

```
The learning_rate is 2 ** -11.5
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 302 and 698 out of 1000.
The time was 00:00:03.6982401 for: iteration 0
Starting timing for: iteration 1
Winrate is 441 and 559 out of 1000.
The time was 00:00:04.8093117 for: iteration 1
Starting timing for: iteration 2
Winrate is 527 and 473 out of 1000.
The time was 00:00:05.3505430 for: iteration 2
Starting timing for: iteration 3
Winrate is 529 and 471 out of 1000.
The time was 00:00:05.5603662 for: iteration 3
Starting timing for: iteration 4
Winrate is 569 and 431 out of 1000.
The time was 00:00:05.6384007 for: iteration 4
Starting timing for: iteration 5
Winrate is 566 and 434 out of 1000.
The time was 00:00:05.5209182 for: iteration 5
Starting timing for: iteration 6
Winrate is 612 and 388 out of 1000.
The time was 00:00:05.7409772 for: iteration 6
Starting timing for: iteration 7
Winrate is 625 and 375 out of 1000.
The time was 00:00:06.5091978 for: iteration 7
Starting timing for: iteration 8
Winrate is 653 and 347 out of 1000.
The time was 00:00:06.2124761 for: iteration 8
Starting timing for: iteration 9
Winrate is 631 and 369 out of 1000.
The time was 00:00:06.6030101 for: iteration 9
Starting timing for: iteration 10
Winrate is 684 and 316 out of 1000.
The time was 00:00:07.2228640 for: iteration 10
Starting timing for: iteration 11
Winrate is 685 and 315 out of 1000.
The time was 00:00:06.9939983 for: iteration 11
Starting timing for: iteration 12
Winrate is 676 and 324 out of 1000.
The time was 00:00:06.8809761 for: iteration 12
Starting timing for: iteration 13
Winrate is 700 and 300 out of 1000.
The time was 00:00:06.6628207 for: iteration 13
Starting timing for: iteration 14
Winrate is 698 and 302 out of 1000.
The time was 00:00:06.6241007 for: iteration 14
```

No, but rather than dividing by primal, I could get more mileage of out simply taking the log instead."