"11:15am. Let me go back to that great asymptotic run. I do not think I should strain the log after all.
11:20am.

```
        inm _ = abs (abs (primal error) - scale) |> as_cost
        inm scale = max epsilon scale

        inm _ = sqr error |> as_cost
```

Without that primal error it works much worse. This is quite interesting. Let me remove the sqr error part.

```
        inm _ = abs (abs error - scale) |> as_cost
        inm scale = max epsilon scale

        //inm _ = sqr error |> as_cost
```

I wonder what this will bring.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 318 and 682 out of 1000.
The time was 00:00:04.0384369 for: iteration 0
Starting timing for: iteration 1
Winrate is 576 and 424 out of 1000.
The time was 00:00:05.6007526 for: iteration 1
Starting timing for: iteration 2
Winrate is 596 and 404 out of 1000.
The time was 00:00:05.5303220 for: iteration 2
Starting timing for: iteration 3
Winrate is 651 and 349 out of 1000.
The time was 00:00:05.5423447 for: iteration 3
Starting timing for: iteration 4
Winrate is 617 and 383 out of 1000.
The time was 00:00:05.4386422 for: iteration 4
Starting timing for: iteration 5
Winrate is 630 and 370 out of 1000.
The time was 00:00:05.0515549 for: iteration 5
Starting timing for: iteration 6
Winrate is 632 and 368 out of 1000.
The time was 00:00:05.2289820 for: iteration 6
Starting timing for: iteration 7
```

Nothing good. I might have discovered something great by accident though.

```
        inm _ = sqr (sqr error - scale) |> as_cost
        inm scale = max epsilon (sqrt scale)
```

Let me try this.

```
        inm _ = sqr (sqr error - scale) |> as_cost
        inm scale = max epsilon (sqrt (abs scale))
```

It just dies when I try to optimize it like this. As one would expect.

11:25am.

```
        inm _ = abs (abs error - scale) |> as_cost
        inm scale = max epsilon scale

        inm _ = sqr error |> as_cost
        inm scaled_error = error / scale // Is used as reward for the actor.
```

I think I will use this from here on out.

Let me try dividing the rewards by 10 so I can see what will happen.

11:30am. The repeated crashes have messed up the GPU. I need to reset.

Nevermind, I'll resume this later. It is time for a break."