"1:15pm.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 447 and 553 out of 1000.
The time was 00:00:05.9179284 for: iteration 0
Starting timing for: iteration 1
Winrate is 654 and 346 out of 1000.
The time was 00:00:06.3873760 for: iteration 1
Starting timing for: iteration 2
Winrate is 689 and 311 out of 1000.
The time was 00:00:06.4687007 for: iteration 2
Starting timing for: iteration 3
Winrate is 719 and 281 out of 1000.
The time was 00:00:07.0158344 for: iteration 3
Starting timing for: iteration 4
Winrate is 711 and 289 out of 1000.
The time was 00:00:07.3837558 for: iteration 4
Starting timing for: iteration 5
Winrate is 706 and 294 out of 1000.
The time was 00:00:06.7152408 for: iteration 5
Starting timing for: iteration 6
Winrate is 749 and 251 out of 1000.
The time was 00:00:06.1843150 for: iteration 6
Starting timing for: iteration 7
Winrate is 736 and 264 out of 1000.
The time was 00:00:06.4422171 for: iteration 7
Starting timing for: iteration 8
Winrate is 710 and 290 out of 1000.
The time was 00:00:06.3268239 for: iteration 8
Starting timing for: iteration 9
Winrate is 718 and 282 out of 1000.
The time was 00:00:06.4337649 for: iteration 9
Starting timing for: iteration 10
Winrate is 745 and 255 out of 1000.
The time was 00:00:06.6101016 for: iteration 10
Starting timing for: iteration 11
Winrate is 706 and 294 out of 1000.
The time was 00:00:06.8214631 for: iteration 11
Starting timing for: iteration 12
Winrate is 724 and 276 out of 1000.
The time was 00:00:06.7726119 for: iteration 12
Starting timing for: iteration 13
Winrate is 720 and 280 out of 1000.
The time was 00:00:06.8433846 for: iteration 13
Starting timing for: iteration 14
Winrate is 723 and 277 out of 1000.
The time was 00:00:06.9658804 for: iteration 14
```

Ah, I forgot to reduce the epsilon. Let me try it again.

1:20pm.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 419 and 581 out of 1000.
The time was 00:00:05.9803416 for: iteration 0
Starting timing for: iteration 1
Winrate is 557 and 443 out of 1000.
The time was 00:00:06.3972579 for: iteration 1
Starting timing for: iteration 2
Winrate is 653 and 347 out of 1000.
The time was 00:00:06.4192959 for: iteration 2
Starting timing for: iteration 3
Winrate is 687 and 313 out of 1000.
The time was 00:00:06.8176370 for: iteration 3
Starting timing for: iteration 4
Winrate is 669 and 331 out of 1000.
The time was 00:00:06.8090873 for: iteration 4
Starting timing for: iteration 5
Winrate is 682 and 318 out of 1000.
The time was 00:00:06.6789412 for: iteration 5
Starting timing for: iteration 6
Winrate is 693 and 307 out of 1000.
The time was 00:00:06.9511998 for: iteration 6
Starting timing for: iteration 7
Winrate is 735 and 265 out of 1000.
The time was 00:00:06.5354910 for: iteration 7
Starting timing for: iteration 8
Winrate is 691 and 309 out of 1000.
The time was 00:00:06.6418816 for: iteration 8
Starting timing for: iteration 9
Winrate is 699 and 301 out of 1000.
The time was 00:00:06.0643027 for: iteration 9
Starting timing for: iteration 10
Winrate is 710 and 290 out of 1000.
The time was 00:00:05.9292570 for: iteration 10
Starting timing for: iteration 11
Winrate is 710 and 290 out of 1000.
The time was 00:00:06.3182502 for: iteration 11
Starting timing for: iteration 12
Winrate is 726 and 274 out of 1000.
The time was 00:00:06.6208928 for: iteration 12
Starting timing for: iteration 13
Winrate is 717 and 283 out of 1000.
The time was 00:00:06.5288808 for: iteration 13
Starting timing for: iteration 14
Winrate is 743 and 257 out of 1000.
The time was 00:00:06.4598611 for: iteration 14
```

Let me also divide the scale."