"1:20pm.
```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 444 and 556 out of 1000.
The time was 00:00:06.4182365 for: iteration 0
Starting timing for: iteration 1
Winrate is 645 and 355 out of 1000.
The time was 00:00:07.2657928 for: iteration 1
Starting timing for: iteration 2
Winrate is 650 and 350 out of 1000.
The time was 00:00:07.7607733 for: iteration 2
Starting timing for: iteration 3
Winrate is 610 and 390 out of 1000.
The time was 00:00:06.1896406 for: iteration 3
Starting timing for: iteration 4
Winrate is 616 and 384 out of 1000.
The time was 00:00:06.3089740 for: iteration 4
Starting timing for: iteration 5
Winrate is 691 and 309 out of 1000.
The time was 00:00:06.4290832 for: iteration 5
Starting timing for: iteration 6
Winrate is 669 and 331 out of 1000.
The time was 00:00:06.3351390 for: iteration 6
Starting timing for: iteration 7
Winrate is 692 and 308 out of 1000.
The time was 00:00:06.0735915 for: iteration 7
Starting timing for: iteration 8
Winrate is 655 and 345 out of 1000.
The time was 00:00:06.2718639 for: iteration 8
Starting timing for: iteration 9
Winrate is 665 and 335 out of 1000.
The time was 00:00:05.9439313 for: iteration 9
Starting timing for: iteration 10
Winrate is 706 and 294 out of 1000.
The time was 00:00:05.9229001 for: iteration 10
Starting timing for: iteration 11
Winrate is 713 and 287 out of 1000.
The time was 00:00:06.4581333 for: iteration 11
Starting timing for: iteration 12
Winrate is 711 and 289 out of 1000.
The time was 00:00:06.3232680 for: iteration 12
Starting timing for: iteration 13
Winrate is 696 and 304 out of 1000.
The time was 00:00:06.4347289 for: iteration 13
Starting timing for: iteration 14
Winrate is 710 and 290 out of 1000.
The time was 00:00:06.2747116 for: iteration 14
```

I am disgusted by this. Literally nothing is going according to plan.

I make an idea, drum up interest to implement it and then reality smacks me down.

I do not want to spend time tuning learning rates and that bullshit. It has to work well on its own!

This is all backprop's fault. The global way of propagating errors is leading me down empty alleyways. It would be so much easier if individual modules had some sense of right and wrong.

Well, let me try pushing up the LR for this and then I will wipe my hands clean of this."