I am going to take a break here. It feels like I've essentially wasted 2 weeks going nowhere with this.
There is so much I can implement, but so little time...

I will forget poker for a while so I can investigate plastic RNNs properly in the context of the Maze task. Given the results of all of this, maybe it is seriously time to consider to just get rid of backpropagation and replace it with something more sensible.

Each neuron making a prediction does make sense. Making these hideous global updates and without even having good pruning is what is causing all this variance.