"4:10pm. I am back from doing chores. Let me finally read the Q-AGREL paper.
4:30pm. "An unexpected observation was that learning proceeded in a stepwise manner for Q-AGREL, which contrasts with the smoother progression with standard error-backpropagation. When we analyzed this result, we found that the network discovered each of the 10 digit classes at a time. When it happened to select a new class during the presentation of a digit of the same class (i.e. a coincidental correct response) as the result of the random action selection policy, it quickly started to recognize most samples of the same class."

...

"I've tried various sort of things to make the univariate Gaussian KL divergence work along with some other ideas, but in the end, at their best, they all work either on par with squared error or slightly worse. What a waste of time. I am quite disappointed by this.

I've tried all sorts of settings, though not different random seeds as the task I am testing it on is quite consistent with KFAC (though not SGD.)

The KL divergence in particular has a tendency to blow up towards the end of training as the scale parameter grows without bound after the value error shrinks. It did not help the reward scaling issue.

The main issue that I've tried to tackle with this experiment, namely the reward scaling remains untouched. Even with KFAC which is supposed to whitening the gradients, with a RNN as the hidden layer the net is still greatly affected by reward scaling.

Now the poker game I am trying it out on is simple, the all the rewards are in the `[-10,10]` range, but the full game would have a far wider range. Given how much trouble `[-10, 10]` is giving it, imagine trying to tackle tourneys which have discrete rewards spread in the `[-1m,1m]` range. It is not that I want an agent that would play those variance fests, but what I do want is an algorithm that would be capable of dealing with these in theory before I am willing to place my trust in RL.

For a toy poker game like the one I am testing the basic prerequisite to meet would be a 100% success rate regardless of reward scaling with deep nets and without needing KFAC to grease its wheels at every step. There is so much KFAC cannot deal with and in fact the reason why scaling rewards is not equal to modifying learning rates is because of the multiplier effect in the critic introduced by it.

Now, I do not think that anticipating the scale is wrong based on the result of this, but it is making me more open to consider alternatives to backprop.

The learnable eligibility trace does provide a slight benefit on the toy poker task and might be worth investigating further.

That concludes my report."

6:05pm. "> An unexpected observation was that learning proceeded in a stepwise manner for Q-AGREL, which contrasts with the smoother progression with standard error-backpropagation. When we analyzed this result, we found that the network discovered each of the 10 digit classes at a time. When it happened to select a new class during the presentation of a digit of the same class (i.e. a coincidental correct response) as the result of the random action selection policy, it quickly started to recognize most samples of the same class.

I feel papers like this one make too big of a deal about biological plausibility as if it is some intrinsic good without elaborating why it is important. I think that reducing learning to local rules could unlock novel metalearning approaches. Something being local implies that it is decoupled and that is something that is necessary for composability. Meaning you could have nets controlling what would otherwise be hyperparameters of other nets.

Besides composability, good local rules should provide stability that backprop lacks except in tightly controlled supervised learning contexts. I would hesitate to call backprop a stable algorithm, and a better one would derive global stability stepwise from local stability.

It might be interesting to evaluate this algorithm on actual RL tasks rather than Mnist."

I am done with lunch. I will call it a day here as I am dead tired. When I said I was going to take a break earlier, I meant it literally rather than having to work around the house. It along with the the recent disappointments has worn me out.

I am considering that replacing backprop might be worth a serious try now. I have no opinion whether Q-AGREL is the one that I want, but I am already moving in the direction of putting Hebbian learning everywhere. I might as well double and triple up on metalearning since I have no better option.

I have no better option than to reconsider the basics of how credit assignment is done in deep nets.

I was hoping I would not have to deal with this, but I am going to have to. If I cannot stabilize the normal nets on this toy task, there is no way the metalearning with plastic RNNs will go the way I want it to.

That unless I attain perfect balance I won't be able to win is something I knew 3.5 years ago.

I have to keep going. It does not matter how many times I fail, I only need to win once."